Remove JCenter repository
This replaces JCenter (which is being sunsetted) with Maven Central and the Gradle Plugins and Releases repositories.

The NoMagic and internal repositories are now placed last to avoid hitting NoMagic's server for every dependency and
to work around a Gradle bug causing a failure when pulling from the internal repository.