Remove overwrite parameter added for NOMAGIC-1205
While reviewing the 4.4.3 release branch a problem was identified with sending the overwrite parameter
when the element being sent is a view, as doing so may remove generated view information.

This reverts the change for the time being to allow the release to go forward until we can work out
how to better handle this situation.